Cards Now Properly Reset
Card will properly reset their position and transform parent when they are dragged to a zone that doesn't exist. Unfortunately they don't retain their original position and instead go to the rightmost part of the screen/hand. Fixed number indicator showing the wrong quantity in the resource consumer. Number indicator is now SetActive: False when in the hand.