Choice Selection works
Selecting a choice will, if it's ready, remove all cards in the consumer and add the rewards to the player's hand. On the to-do list is currently to add function comments to every function in the Game Manager. This will let me easier look back and see the purpose of every function.