Not sure what is happening
The borderColorOverlay is changing colors for other instances of prefabs when changing a single instance's color. This is happening with different prefabs, which is the strangest part. The color change is only present in half of the card's border and only present in the scene view; not the game view.